Rename remaps file
